Prevents crash when searching
Adds a try catch for when preventing a crash when searching. This may or may not be a permanent fix. When checking the newest version of other map projects using the same framework the console still reports this error but it does not crash. They do not use this fix but they must use something similar.