Better clustering integration.
Clustering is now usefull. I am using the fossil grouping as a base but mindat has a lot more points so it will need to be refactored to make more sense. Basically the design of clustered points needs to be redesigned to make more sense. Also need to fix issue where mindat points stick around after we zoom out to clusters.