Points and clusters move with age.
Points and custers move with age. Some temperamental features such as mapbox bounds after changing, not sure I can do much though. Some weird cluster locations show up and I believe it has to do with a mismatch between the size of the cluster points and the size of the cluster radius.